requirements.txt
streamlit
mysql-connector-python
pillow
pandas
matplotlib
